Logger changes, configuration manager changes
Added unit tests for configuration manager, adapted tests for logger,
changed config file structure

+++ first_semester/programing/final_project/murrent/BiOWheels/BiOWheelsLogger/CombinedLogger.cs
@@ -25,7 +25,7 @@
         }
 
         /// <inheritdoc/>
-        public void SetFileSize<T>(long logFileSize)
+        public void SetFileSize<T>(double logFileSize)
         {
             if (typeof(T) == typeof(FileLogger))
             {
